Redo monospace alphabet
The XITS version I used earlier seems to be based on cmtt10, so I
modified the METAFONT parameters of the later to get the changes I need
and used it instead. The modified parameters are:

    asc_height#:=245/36pt#;
    cap_height#:=232/36pt#;
    fig_height#:=215/36pt#;
    desc_depth#:=85/36pt#;

Used mftrace to trace the MATFONT source, so the curves are a bit
suboptimal. Tried mf2pt1 but it produced garbage, and mflua needs a PhD
to figure out how to run it.