Fixes #90 Historic ligatures
Corrects the `c_t` and `s_t` ligature classification across all the fonts of the family.

Fixes https://github.com/khaledhosny/libertinus/issues/90

Summary:
*) Serif Semibold-italic:
- renamed from `V.alt` to `c_t` the glyph actually containing the `c_t` ligature
- added the `hlig` table
- added `c_t` to the `hlig` table
- moved `s_t` from the `dlig` table to the `hlig` table
- removed the empty `dlig` table

*) Serif Bold-italic:
- added the missing `s_t` ligature glyph
- added `s_t` to the `hlig` table
- removed the empty `dlig` table

In the Serif Semibold and Bold, the `c_t` ligature is different from the other styles and more 'flat': I have not changed it, for the sake of variety.