Fix #162 : Names of A/O/Udieresis and A/O/Udieresis.alt
Fixes https://github.com/libertinus-fonts/libertinus/issues/162

Glyphs U+00C4, U+00D6 and U+00DC with the regular shapes for A/O/U+dieresis are named `Adieresis.alt`, `Odieresis.alt` and `Udieresis.alt`, while alternate glyphs with lower dieresis in unencoded slots are named `Adieresis`, Odieresis` and `Udieresis`.

This PR swaps the names for a more consistent naming.

The Serif Bold-italic was different, as the names were correct but the shapes were exchanged; for this font, the PR swaps the shapes rather than the names.