Cleanup whitespace in Makefile, good for verbose console output
One could argue about the style of extra whitespace before line
continuation, but I would argue the output formatting is as important as
the input formating. One thing it makes really messy is the console
output of echoing the rules as they are run. In silent mode (the
default) this may not matter but for debugging purposes (and probably CI
later) its useful to be able to follow along with whats actually being
run. This makes that a lot easier by not introducing a bunch of
whitespace that breaks wrapping in the output.