Drop generated binary fonts from repository
With all due respect to previous maintainers (and indeed the vast
majority of designers dumping their fonts into Git) this is and always
was an incorrect use of version control. The problems with it are
legion: massively inflated repository size, compete inability to use
VCS tooling for things like diff, merge, rebase, revert, etc. The
two main "problems" this work flow solves can easily be solved with the
proper tools for the job: reproducible build scripts and archives of the
generated fonts.

This font project in particular is in an excellent position to take this
leap because the builds are already reproducible and the primary venue
for releases (Github) allows attaching archives to releases.