Add (back) preview specimen regeneration
Turns out this hasn't been working for a long time. I did some `git
blame` and found at one point my own contribution broke it (used the
make target as input file, input as output!), but it turns out it was
broken before that too. Or `mutool` was different in those days or
something.

Since I plan to replace this with a SILE specimen anyway I'm not too
worried, but this gets it back to where it purported to being first.