chore(tooling): Relax requirement on updated preview for the sake of CI
This didn't sed to be a problem because of a race condition in Fontship
that resulted in the rule being discarded anyway. When we fixed that bug
(using double expansion and a POSTFONTSHIPEVAL) we introduced
a dependency since we don't have XeLaTeX available in CI. This makes it
possible to regenerate locally but not required to generate packages.