refactor(tooling): Swap out hack job to change the sample background for an actual SVG tool
Mutool changed its output, so our `sed` trick stopped working to replace
the background. I started futzing with it more and it turns out the
output is radically different, so I just went ahead and used real SVG
tooling. First to add the background and resize to match the old preview
(so tools like `lukaj` can preview the difference) then to compress and
normalize a bit.

I opted against updating the output because it's 1/2 MB of content with
no meaningful changes for this release. We changed lots of stuff in
other languages, but this particular preview is visually identical even
if we re-generate it with the new encodings.